Holdのコンストラクタを追加、その他移植
