+++ src/Jmemo2/YoubeatParser.cpp
@@ -50,7 +50,7 @@ std::vector<jubeon::game::Note> jubeon::parser::YoubeatParser::parse(const std::
 	//Prepare BPM Table
 	BpmTable bpm_table;
 	
-	std::list<
+	//std::list<
 
 	//parse block
 	for (auto line = lines.begin(); line != lines.end(); line++) {
@@ -237,32 +237,3 @@ jubeon::parser::YoubeatParser::BpmColumn::BpmColumn(const unsigned int haku, con
 }
 
 
-jubeon::parser::YoubeatParser::Note_C::Note_C(const unsigned int haku, const unsigned int m, const unsigned int v, const jubeon::jPanel p_no)
-	: haku(haku), m(m), v(v),p_no(p_no)
-{
-}
-
-jubeon::game::Note jubeon::parser::YoubeatParser::Note_C::convertToNote(const BpmTable bpm_table, const jMillisec offset) const
-{
-	BpmTable::const_iterator p = bpm_table.begin();
-
-	for (; p != bpm_table.end(); p++) {
-		if (p->haku > this->haku) break;
-	}
-
-	double paramA = 0.0f;
-	double paramB = 0.0f;
-	double ms = 0.0f;
-
-	// != p ‚Ü‚Å‚Æ‚µ‚ÄŒvZ‚·‚ê‚Î‚¢‚¢
-	for (BpmTable::const_iterator cur = bpm_table.begin(); cur != p; cur++) {
-		ms = paramA * cur->haku + paramB;
-		
-		paramA = 60000.0f / cur->bpm;
-		paramB = ms - paramA * cur->haku;
-	}
-
-	//ms = paramA * (haku + m / v) + paramB
-	return jubeon::game::Note(paramA * (static_cast<double>(this->haku) + static_cast<double>(this->m) / static_cast<double>(this->v)) + paramB + offset, this->p_no);
-
-}
